These changes correct structure assumptions for APPEARANCE_PC and COMMANDER. In turn, the packet ZC_CONNECT_OK is fixed which allows for connection to the zone server.

+++ src/Shared/World/BaseCharacter.cs
@@ -95,12 +95,12 @@ namespace Melia.Shared.World
 		/// <summary>
 		/// Spell points.
 		/// </summary>
-		public int Sp { get; set; }
+		public short Sp { get; set; }
 
 		/// <summary>
 		/// Maximum spell points.
 		/// </summary>
-		public int MaxSp { get; set; }
+		public short MaxSp { get; set; }
 
 		/// <summary>
 		/// Stamina points.
